requirements.txt
Django
ipython
pyodbc
django-admin-tools
callfire-api-client-python
python-dotenv